test : 객체 역할 분리에 따른 테스트 코드 리팩토링
